requirements.txt
# MCP Core
mcp>=0.1.0

# PostgreSQL
psycopg2-binary>=2.9.0
asyncpg>=0.29.0

# System Monitoring
psutil>=5.9.0

# HTTP Clients
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# GitHub Integration
PyGithub>=2.1.0

# Embeddings (for memory MCP)
sentence-transformers>=2.2.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0